feat(user): - Added UpdatePassword use case to handle password updates with session revocation option.
chore(user): refactor user-related DTOs and use cases for improved structure and functionality

- Consolidated authentication-related DTOs into a single auth.dto.ts file.
- Updated user DTOs to enhance type safety and validation.
- Refactored user.controller.ts to include updatePassword method and corresponding routes.
- Adjusted validation schemas for user updates and password changes.
- Improved logging and error handling in the UpdatePassword use case.
- Updated app.router.ts and user.router.ts to reflect new password update route.

Issue T103
Close #30
