+++ src/application/index.ts
@@ -1,9 +1,7 @@
 // Dto's
-export * from './dtos/auth/code.dto.js';
-export * from './dtos/auth/login.dto.js';
-export * from './dtos/auth/token.dto.js';
-export * from './dtos/client/validate-client.dto.js';
-export * from './dtos/user/user.dto.js';
+export * from './dtos/auth.dto.js';
+export * from './dtos/validate-client.dto.js';
+export * from './dtos/user.dto.js';
 
 //Use cases
 export * from './use-cases/auth/exchange-token.use-case.js';
@@ -15,6 +13,7 @@ export * from './use-cases/client/validate-client.use-case.js';
 export * from './use-cases/user/register-user.use-case.js';
 export * from './use-cases/user/get-user.use-case.js';
 export * from './use-cases/user/update-user.use-case.js';
+export * from './use-cases/user/update-password.use-case.js';
 
 //Validate Schemas
 export * from './validate-schemas/auth.schemas.js';
